requirements.txt
PostgreSQL connector:
psycopg2 
djangorestframework
django-rest-auth
django-rest-framework-jwt
